Implement barebones login function
Add login action to the login button on login screen

sys.path.append had to be used instead of simple import because it
seems that imported file can't import any module from the same
directory or something like that.

Not sure if user.user.User is a correct way to reference the constructor.

Primary key column from 'users' table is actually id_user, not id, so
it's been changed.