Instantiate screen manager in .kv
Change the structure of program by creating objects in kv file. Before,
it was done by adding widgets directly in the tunedune.py file. It
cleared the way objects relate to each other. Hopefully it will
facilitate accessing fields of other classes.

All of this was motivated by inability to connect different objects,
in this case, the main screen and the login screen. It was
necessary to transfer the number of the current session.